`on_auto_moderation_action_execution` event
